feat: typed BookingDetails hierarchy, extra bag, and migration to camelCase JSONB
- Add BookingDetails base interface with source, rawText, extra fields;
  FlightDetails and TransportDetails now extend it; add BaseBookingDetails
  (kind: 'base') for untyped booking types (accommodation, taxi, etc.)
- Add extra?: Record<string, unknown> to all three MCP tool input schemas
  (travel_add_flight, travel_add_transport, travel_add_reservation_from_text)
  so AI can pass additional fields it considers relevant
- Migration 0024: fix JSONB double-encoding (string → object), rename
  snake_case keys to camelCase, add missing kind discriminators
- Extract booking expanded panel into BookingExpandedPanel.tsx; replace
  all four IIFEs with named components (FlightDetailsPill, TransportDetailsPill,
  ExtraDetailsSection, SourceTextSection)
- Show details.extra key/value grid and collapsible details.rawText source
  text in the expanded booking panel
- Fall back to rawText excerpt as secondary label in coming-next timeline
- Replace TripsSidebar IIFE with formatTripDateRange util; replace
  ProfileCard macro warning IIFE with macroSummary?.isOver
